+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/controllers/PermissionController.kt
@@ -1,29 +1,25 @@
-package com.lucasjosino.on_audio_query.controller
+package com.lucasjosino.on_audio_query.controllers
 
 import android.Manifest
-import android.app.Activity
-import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Build
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
+import com.lucasjosino.on_audio_query.PluginProvider
 import com.lucasjosino.on_audio_query.interfaces.PermissionManagerInterface
 import io.flutter.Log
-import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.PluginRegistry
 
-class PermissionController(
-    private var retryRequest: Boolean = false
-) : PermissionManagerInterface, PluginRegistry.RequestPermissionsResultListener {
+class PermissionController : PermissionManagerInterface,
+    PluginRegistry.RequestPermissionsResultListener {
 
     companion object {
         private const val TAG: String = "PermissionController"
+
         private const val REQUEST_CODE: Int = 88560
     }
 
-    private lateinit var activity: Activity
-
-    private lateinit var result: MethodChannel.Result
+    var retryRequest: Boolean = false
 
     private var permissions: Array<String> =
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
@@ -38,33 +34,31 @@ class PermissionController(
             )
         }
 
-    override fun permissionStatus(context: Context): Boolean = permissions.all {
+    override fun permissionStatus(): Boolean = permissions.all {
         // After "leaving" this class, context will be null so, we need this context argument to
         // call the [checkSelfPermission].
         return ContextCompat.checkSelfPermission(
-            context,
+            PluginProvider.context(),
             it
         ) == PackageManager.PERMISSION_GRANTED
     }
 
-    override fun requestPermission(activity: Activity, result: MethodChannel.Result) {
+    override fun requestPermission() {
         Log.d(TAG, "Requesting permissions.")
         Log.d(TAG, "SDK: ${Build.VERSION.SDK_INT}, Should retry request: $retryRequest")
-        this.activity = activity
-        this.result = result
+        val activity = PluginProvider.activity()
         ActivityCompat.requestPermissions(activity, permissions, REQUEST_CODE)
     }
 
     // Second requestPermission, this one with the option "Never Ask Again".
     override fun retryRequestPermission() {
+        val activity = PluginProvider.activity()
         if (ActivityCompat.shouldShowRequestPermissionRationale(activity, permissions[0])
             || ActivityCompat.shouldShowRequestPermissionRationale(activity, permissions[1])
         ) {
             Log.d(TAG, "Retrying permission request")
             retryRequest = false
-            if (this::activity.isInitialized && this::result.isInitialized) {
-                requestPermission(activity, result)
-            }
+            requestPermission()
         }
     }
 
@@ -73,11 +67,6 @@ class PermissionController(
         permissions: Array<out String>,
         grantResults: IntArray
     ): Boolean {
-        // When [pResult] is not initialized the permission request did not originate from the
-        // [on_audio_query] plugin, so return [false] to indicate the [on_audio_query] plugin is not
-        // handling the request result and Android should continue executing other registered handlers.
-        if (!this::result.isInitialized) return false
-
         // When the incoming request code doesn't match the request codes defined by the on_audio_query
         // plugin return [false] to indicate the [on_audio_query] plugin is not handling the request
         // result and Android should continue executing other registered handlers.
@@ -90,6 +79,7 @@ class PermissionController(
         Log.d(TAG, "Permission accepted: $isPermissionGranted")
 
         // After all checks, we can handle the permission request.
+        val result = PluginProvider.result()
         when {
             isPermissionGranted -> result.success(true)
             retryRequest -> retryRequestPermission()