+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/queries/AlbumQuery.kt
@@ -1,17 +1,15 @@
 package com.lucasjosino.on_audio_query.queries
 
 import android.content.ContentResolver
-import android.content.Context
 import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.controller.PermissionController
+import com.lucasjosino.on_audio_query.PluginProvider
+import com.lucasjosino.on_audio_query.controllers.PermissionController
 import com.lucasjosino.on_audio_query.queries.helper.QueryHelper
 import com.lucasjosino.on_audio_query.types.checkAlbumsUriType
 import com.lucasjosino.on_audio_query.types.sorttypes.checkAlbumSortType
 import io.flutter.Log
-import io.flutter.plugin.common.MethodCall
-import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -32,17 +30,11 @@ class AlbumQuery : ViewModel() {
 
     /**
      * Method to "query" all albums.
-     *
-     * Parameters:
-     *   * [context]
-     *   * [result]
-     *   * [call]
      */
-    fun queryAlbums(
-        context: Context,
-        result: MethodChannel.Result,
-        call: MethodCall
-    ) {
+    fun queryAlbums() {
+        val call = PluginProvider.call()
+        val result = PluginProvider.result()
+        val context = PluginProvider.context()
         this.resolver = context.contentResolver
 
         // Sort: Type and Order.
@@ -62,7 +54,7 @@ class AlbumQuery : ViewModel() {
         Log.d(TAG, "\turi: $uri")
 
         // We cannot 'query' without permission.
-        val hasPermission: Boolean = PermissionController().permissionStatus(context)
+        val hasPermission: Boolean = PermissionController().permissionStatus()
         if (!hasPermission) {
             result.error(
                 "403",
@@ -74,7 +66,7 @@ class AlbumQuery : ViewModel() {
 
         // Query everything in background for a better performance.
         viewModelScope.launch {
-            val queryResult: ArrayList<MutableMap<String, Any?>> = loadAlbums()
+            val queryResult = loadAlbums()
             result.success(queryResult)
         }
     }