+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/queries/ArtistQuery.kt
@@ -1,18 +1,16 @@
 package com.lucasjosino.on_audio_query.queries
 
 import android.content.ContentResolver
-import android.content.Context
 import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.controller.PermissionController
+import com.lucasjosino.on_audio_query.PluginProvider
+import com.lucasjosino.on_audio_query.controllers.PermissionController
 import com.lucasjosino.on_audio_query.queries.helper.QueryHelper
 import com.lucasjosino.on_audio_query.types.checkArtistsUriType
 import com.lucasjosino.on_audio_query.types.sorttypes.checkArtistSortType
 import com.lucasjosino.on_audio_query.utils.artistProjection
 import io.flutter.Log
-import io.flutter.plugin.common.MethodCall
-import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -34,14 +32,12 @@ class ArtistQuery : ViewModel() {
 
     /**
      * Method to "query" all artists.
-     *
-     * Parameters:
-     *   * [context]
-     *   * [result]
-     *   * [call]
      */
-    fun queryArtists(context: Context, result: MethodChannel.Result, call: MethodCall) {
-        resolver = context.contentResolver
+    fun queryArtists() {
+        val call = PluginProvider.call()
+        val result = PluginProvider.result()
+        val context = PluginProvider.context()
+        this.resolver = context.contentResolver
 
         // Sort: Type and Order
         sortType = checkArtistSortType(
@@ -60,7 +56,7 @@ class ArtistQuery : ViewModel() {
         Log.d(TAG, "\turi: $uri")
 
         // We cannot 'query' without permission.
-        val hasPermission: Boolean = PermissionController().permissionStatus(context)
+        val hasPermission: Boolean = PermissionController().permissionStatus()
         if (!hasPermission) {
             result.error(
                 "403",
@@ -72,7 +68,7 @@ class ArtistQuery : ViewModel() {
 
         // Query everything in background for a better performance.
         viewModelScope.launch {
-            val queryResult: ArrayList<MutableMap<String, Any?>> = loadArtists()
+            val queryResult = loadArtists()
             result.success(queryResult)
         }
     }