+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/queries/GenreQuery.kt
@@ -1,18 +1,16 @@
 package com.lucasjosino.on_audio_query.queries
 
 import android.content.ContentResolver
-import android.content.Context
 import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.controller.PermissionController
+import com.lucasjosino.on_audio_query.PluginProvider
+import com.lucasjosino.on_audio_query.controllers.PermissionController
 import com.lucasjosino.on_audio_query.queries.helper.QueryHelper
 import com.lucasjosino.on_audio_query.types.checkGenresUriType
 import com.lucasjosino.on_audio_query.types.sorttypes.checkGenreSortType
 import com.lucasjosino.on_audio_query.utils.genreProjection
 import io.flutter.Log
-import io.flutter.plugin.common.MethodCall
-import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -33,14 +31,12 @@ class GenreQuery : ViewModel() {
 
     /**
      * Method to "query" all genres.
-     *
-     * Parameters:
-     *   * [context]
-     *   * [result]
-     *   * [call]
      */
-    fun queryGenres(context: Context, result: MethodChannel.Result, call: MethodCall) {
-        resolver = context.contentResolver
+    fun queryGenres() {
+        val call = PluginProvider.call()
+        val result = PluginProvider.result()
+        val context = PluginProvider.context()
+        this.resolver = context.contentResolver
 
         // Sort: Type and Order.
         sortType = checkGenreSortType(
@@ -59,7 +55,7 @@ class GenreQuery : ViewModel() {
         Log.d(TAG, "\turi: $uri")
 
         // We cannot 'query' without permission.
-        val hasPermission: Boolean = PermissionController().permissionStatus(context)
+        val hasPermission: Boolean = PermissionController().permissionStatus()
         if (!hasPermission) {
             result.error(
                 "403",
@@ -71,7 +67,7 @@ class GenreQuery : ViewModel() {
 
         // Query everything in background for a better performance.
         viewModelScope.launch {
-            val queryResult: ArrayList<MutableMap<String, Any?>> = loadGenres()
+            val queryResult = loadGenres()
             result.success(queryResult)
         }
     }