+++ on_audio_query/ios/Classes/controllers/MethodController.swift
@@ -1,34 +1,29 @@
 import Flutter
 
 public class MethodController {
-    private var call: FlutterMethodCall
-    private var result: FlutterResult
-
-    init(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
-        self.call = call
-        self.result = result
-    }
-
     public func find() {
+        let call = try! PluginProvider.call()
+        let result = try! PluginProvider.result()
+
         // All necessary method to this plugin support both platforms, only playlists
         // are limited when using [IOS].
         switch call.method {
         case Method.QUERY_AUDIOS:
-            AudioQuery(call, result).querySongs()
+            AudioQuery().querySongs()
         case Method.QUERY_ALBUMS:
-            AlbumQuery(call, result).queryAlbums()
+            AlbumQuery().queryAlbums()
         case Method.QUERY_ARTISTS:
-            ArtistQuery(call, result).queryArtists()
+            ArtistQuery().queryArtists()
         case Method.QUERY_GENRES:
-            GenreQuery(call, result).queryGenres()
+            GenreQuery().queryGenres()
         case Method.QUERY_PLAYLISTS:
-            PlaylistQuery(call, result).queryPlaylists()
+            PlaylistQuery().queryPlaylists()
         case Method.QUERY_AUDIOS_FROM:
-            AudioFromQuery(call, result).queryAudiosFrom()
+            AudioFromQuery().queryAudiosFrom()
         case Method.QUERY_WITH_FILTERS:
-            WithFiltersQuery(call, result).queryWithFilters()
+            WithFiltersQuery().queryWithFilters()
         case Method.QUERY_ARTWORK:
-            ArtworkQuery(call, result).queryArtwork()
+            ArtworkQuery().queryArtwork()
         // The playlist for [IOS] is completely limited, the developer can only:
         //   * Create playlist
         //   * Add item to playlist (Unsuported, for now)
@@ -39,9 +34,9 @@ public class MethodController {
         //   * Remove item from playlist
         //   * Move item inside playlist
         case Method.CREATE_PLAYLIST:
-            PlaylistController(call, result).createPlaylist()
+            PlaylistController().createPlaylist()
         case Method.ADD_TO_PLAYLIST:
-            PlaylistController(call, result).addToPlaylist()
+            PlaylistController().addToPlaylist()
         default:
             // All non suported methods will throw this error.
             result(FlutterMethodNotImplemented)