+++ on_audio_query/ios/Classes/controllers/PlaylistController.swift
@@ -5,9 +5,9 @@ class PlaylistController {
     var args: [String: Any]
     var result: FlutterResult
     
-    init(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
-        self.args = call.arguments as! [String: Any]
-        self.result = result
+    init() {
+        self.args = try! PluginProvider.call().arguments as! [String: Any]
+        self.result = try! PluginProvider.result()
     }
     
     func createPlaylist() {